perbaikan PMD TEST

+++ src/main/java/id/ac/ui/cs/advprog/eshop/service/ProductServiceImpl.java
@@ -2,7 +2,6 @@ package id.ac.ui.cs.advprog.eshop.service;
 
 import id.ac.ui.cs.advprog.eshop.model.Product;
 import id.ac.ui.cs.advprog.eshop.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -12,8 +11,12 @@ import java.util.List;
 @Service
 public class ProductServiceImpl implements ProductService {
 
-    @Autowired
-    private ProductRepository productRepository;
+    private final ProductRepository productRepository;
+
+
+    public ProductServiceImpl(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
 
     @Override
     public Product create(Product product) {
@@ -35,6 +38,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void update(Product updatedProduct) {
+        if (updatedProduct == null || updatedProduct.getProductId() == null) {
+            return;
+        }
+
         Product existingProduct = productRepository.findById(updatedProduct.getProductId());
         if (existingProduct != null) {
             existingProduct.setProductName(updatedProduct.getProductName());
@@ -44,6 +51,8 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteById(String id) {
-        productRepository.deleteById(id);
+        if (id != null) {
+            productRepository.deleteById(id);
+        }
     }
 }
